Switch from git submodules to Meson wrapfiles
Meson wrapfiles should be preferred in a Meson project. They make
integration with Meson simpler (no need to specify include dirs or other
things, it's all in the dependency object). They are also use versioned
releases of libraries and not arbitrary commits. Versioned libraries can
be more stable and reliable.